Fix weird protobuf changes (#232)
* feat: add support for git references

* fix: type hierarchy logic that retrieves the package name

it now gets a file URI which confused the logic to search for the next pubspec.yaml to gather the package name.
This ended up in taking the file name as a package name which changed for protobug from "protobuf.dart" to "internal.dart" causing dart_apitool to detect a change.
Additionally package refs ('package:xx') were not treated correctly

* fix: don't treat private top level functions as part of the public API